Refactor: Update imports and fix type errors
