+++ src/algebra/group/__tests__/first-iso.test.ts
@@ -1,5 +1,5 @@
 import { ZmodAdd } from "../examples";
-import { hom as groupHom } from "../../../structures/group/Hom.js";
+import { hom as groupHom } from "../Hom";
 import { analyzeGroupHom } from "../analyzeHom";
 import { firstIsomorphism } from "../FirstIso";
 import { quotientGroup } from "../Quotient";
