+++ src/algebra/group/__tests__/image-kernel-simple.test.ts
@@ -1,7 +1,7 @@
 import { strict as A } from "assert";
 import { imageSubgroup } from "../Image";
 import { kernelNormalSubgroup } from "../Kernel";
-import { hom as groupHom } from "../../../structures/group/Hom.js";
+import { hom as groupHom } from "../Hom";
 import { analyzeGroupHom } from "../analyzeHom";
 import { Cyclic } from "../Group";
